# openapi.yaml
...
openapi: 3.0.3
info:
  title: "GitHub Insight Engine API"
  description: |
    REST API for GitHub data collection, analysis, and insights generation.
    
    **Authentication**: JWT tokens obtained via `/auth/login`
    **Data Flow**: Collect → Store → Analyze → Retrieve
    **Rate Limits**: Standard GitHub API rate limits apply during collection
    **Automatic Scheduling**: Built-in scheduling service with automatic token management
  version: 0.1.90
  contact:
    name: Github
    url: https://github.com/octocat/gh-insight-engine
  license:
    name: MIT
    url: https://github.com/octocat/gh-insight-engine/blob/main/LICENSE
servers:
  - url: https://github.lucas-blanchard.com/api
    description: Production server
  - url: http://localhost:3000/api
    description: Development server
security:
  - BearerAuth: []
  - {}
paths:
...
  /users/{username}:
    post:
      summary: Collect user profile data
      description: Fetch and store GitHub user profile data from GitHub API
      operationId: collectUserData
      tags:
        - Users
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UsernameParam'
      responses:
        '201':
          description: User data collected successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserCollectionResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Error type
                    example: "Unauthorized"
                  message:
                    type: string
                    description: Human-readable error message
                    example: "Please log in to access this resource"
                  timestamp:
                    type: string
                    format: date-time
                    description: Response timestamp
                    example: "2025-07-28T07:36:05.494Z"
        '403':
          description: Can only collect own user data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: GitHub API rate limit exceeded
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Error type
                    example: "RATE_LIMIT_ERROR"
                  message:
                    type: string
                    description: Human-readable error message
                    example: "GitHub API rate limit exceeded. Please try again later or contact GitHub Support if the issue persists."
                  timestamp:
                    type: string
                    format: date-time
                    description: Response timestamp
                    example: "2025-07-28T07:36:05.494Z"

    get:
      summary: Get user profile
      description: Retrieve stored user profile and analysis status
      operationId: getUserProfile
      tags:
        - Users
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UsernameParam'
      responses:
        '200':
          description: User profile retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfileResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Error type
                    example: "Unauthorized"
                  message:
                    type: string
                    description: Human-readable error message
                    example: "Please log in to access this resource"
                  timestamp:
                    type: string
                    format: date-time
                    description: Response timestamp
                    example: "2025-07-28T07:36:05.494Z"
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      summary: Delete user and all associated data
      description: Delete user profile and all associated repositories and AI analyses
      operationId: deleteUserData
      tags:
        - Users
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UsernameParam'
      responses:
        '200':
          description: User and associated data deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDeletionResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Error type
                    example: "Unauthorized"
                  message:
                    type: string
                    description: Human-readable error message
                    example: "Please log in to access this resource"
                  timestamp:
                    type: string
                    format: date-time
                    description: Response timestamp
                    example: "2025-07-28T07:36:05.494Z"
        '403':
          description: Can only delete own user data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token obtained from /auth/login (24h validity).
        Used by both manual API calls and the automatic scheduling service.
  parameters:
    UsernameParam:
      name: username
      in: path
      required: true
      description: GitHub username
      schema:
        type: string
        pattern: '^[a-zA-Z0-9]([a-zA-Z0-9\-]){0,38}$'
        example: "octocat"
  schemas:
...
    UserCollectionResponse:
      type: object
      properties:
        message:
          type: string
          example: "User profile collection successful"
        status:
          type: string
          enum: ["completed", "partial", "failed"]
        userProfile:
          type: object
          properties:
            login:
              type: string
            id:
              type: integer
            name:
              type: string
            public_repos:
              type: integer
            followers:
              type: integer
            following:
              type: integer
            created_at:
              type: string
              format: date-time
        metadata:
          type: object
          properties:
            collectedAt:
              type: string
              format: date-time
            source:
              type: string
              example: "github_api"
        timestamp:
          type: string
          format: date-time
    UserProfileResponse:
      type: object
      properties:
        user:
          type: object
          properties:
            id:
              type: string
            username:
              type: string
            name:
              type: string
            avatarUrl:
              type: string
            bio:
              type: string
              nullable: true
            company:
              type: string
              nullable: true
            location:
              type: string
              nullable: true
            publicRepos:
              type: integer
            followers:
              type: integer
            following:
              type: integer
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
        analysis:
          type: object
          properties:
            hasStoredData:
              type: boolean
            lastAnalyzed:
              type: string
              format: date-time
            hasAiInsights:
              type: boolean
            repositoriesStored:
              type: integer
            dataSource:
              type: string
              enum: ["database", "live", "cache"]
        timestamp:
          type: string
          format: date-time

    # Repository Operations
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          enum:
            - "VALIDATION_ERROR"
            - "AUTHENTICATION_ERROR"
            - "AUTHORIZATION_ERROR"
            - "NOT_FOUND_ERROR"
            - "EXTERNAL_SERVICE_ERROR"
            - "RATE_LIMIT_ERROR"
            - "INTERNAL_SERVER_ERROR"
          example: "VALIDATION_ERROR"
        message:
          type: string
          example: "The provided data does not match the expected format"
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
                example: "githubToken"
              message:
                type: string
                example: "Username is required"
              code:
                type: string
                example: "too_small"
        timestamp:
          type: string
          format: date-time
          example: "2025-07-27T10:18:08.760Z"
        path:
          type: string
          example: "/api/auth/login"
        method:
          type: string
          example: "POST"
        stack:
          type: string
          description: Error stack trace (development only)
...
    UserDeletionResponse:
      type: object
      properties:
        message:
          type: string
          example: "User and associated data deleted successfully"
        status:
          type: string
          enum: ["completed"]
          example: "completed"
        summary:
          type: object
          properties:
            username:
              type: string
              example: "octocat"
            userDeleted:
              type: boolean
              description: Whether the user was successfully deleted
              example: true
            repositoriesDeleted:
              type: integer
              description: Number of repositories deleted
              example: 5
            aiAnalysesDeleted:
              type: integer
              description: Number of AI analyses deleted
              example: 2
            dataFreshness:
              type: string
              enum: ["database"]
              example: "database"
        metadata:
          type: object
          properties:
            deletedAt:
              type: string
              format: date-time
              description: When the deletion occurred
              example: "2025-07-28T07:36:05.494Z"
        timestamp:
          type: string
          format: date-time
          example: "2025-07-28T07:36:05.494Z"
...